Remove non-interactive mode
